add interfaces for services
